+++ src/core/utils/api.js
@@ -1,6 +1,6 @@
 export async function getRollingList() {
   try {
-    const response = await fetch("../constant/headline.json")
+    const response = await fetch("../data/headline.json")
       .then((response) => response.json())
       .then((json) => {
         return json.newsList;
@@ -14,7 +14,7 @@ export async function getRollingList() {
 
 export async function getPressObj() {
   try {
-    const response = await fetch("../constant/pressObj.json")
+    const response = await fetch("../data/pressObj.json")
       .then((response) => response.json())
       .then((json) => {
         return json.pressList;
@@ -28,7 +28,7 @@ export async function getPressObj() {
 
 export async function getNewsContents() {
   try {
-    const response = await fetch("../constant/newsContents.json")
+    const response = await fetch("../data/newsContents.json")
       .then((response) => response.json())
       .then((json) => {
         return json.categoryList;
